Add support to set the GoCD environment correctly
The plugin currently breaks the contract by not respecting the environment GoCD tells it to create an agent in. This can cause agents/instances to spin up that never get assigned jobs, and in some cases agents to get assigned jobs that they should not, due to hacks users use to set the environment with custom user data.

This implements what is needed with a standard GoCD agent to conform to the spec at https://plugin-api.gocd.org/current/elastic-agents/#create-agent